Only trim spaces of array parameter entries
Some packages include a space in the names of their license. In
particular this is a problem with "Public Domain" in org.json:json here:
https://github.com/stleary/JSON-java/blob/master/pom.xml#L32

Add a new function to read an array of strings from an environment
variable that does not remove all spaces but only ones at the beginning
and end of each individual entry. This preserves spaces inside the
entries. Read list of allowed licenses using this new function.