Downgrade Git to version compatible with Bitbucket 8.16.1
Bitbucket 8.16.1 doesn't support Git 2.49.0 and refuses to start with error:
'You are using Git 2.49.0. This version is not supported due to changes
which break core Bitbucket functionality.'

This step downgrades Git to 2.43.x which is compatible with Bitbucket 8.16.1.
If that version isn't available, it will install whatever version is in the
default Ubuntu repos (which should be older and compatible).